+++ lib/eve_online/esi/models/ancestry.rb
@@ -1,0 +1,44 @@
+# frozen_string_literal: true
+
+module EveOnline
+  module ESI
+    module Models
+      class Ancestry < Base
+        def as_json
+          {
+            ancestry_id: ancestry_id,
+            name: name,
+            bloodline_id: bloodline_id,
+            description: description,
+            short_description: short_description,
+            icon_id: icon_id
+          }
+        end
+
+        def ancestry_id
+          options['id']
+        end
+
+        def name
+          options['name']
+        end
+
+        def bloodline_id
+          options['bloodline_id']
+        end
+
+        def description
+          options['description']
+        end
+
+        def short_description
+          options['short_description']
+        end
+
+        def icon_id
+          options['icon_id']
+        end
+      end
+    end
+  end
+end