+++ spec/eve_online/esi/bloodlines_spec.rb
@@ -36,14 +36,14 @@ describe EveOnline::ESI::Bloodlines do
 
     before do
       #
-      # subject.response # => [{"bloodline_id"=>4, "name"=>"Brutor", "description"=>"A martial, strong-willed people, the Brutor...", "race_id"=>2, "ship_type_id"=>588, "corporation_id"=>1000049, "perception"=>9, "willpower"=>7, "charisma"=>6, "memory"=>4, "intelligence"=>4 }]
+      # subject.response # => [{"bloodline_id"=>4, "name"=>"Brutor", "description"=>"A martial, strong-willed people, the Brutor...", "race_id"=>2, "ship_type_id"=>588, "corporation_id"=>1000049, "perception"=>9, "willpower"=>7, "charisma"=>6, "memory"=>4, "intelligence"=>4}]
       #
       expect(subject).to receive(:response).and_return(response)
     end
 
     before do
       #
-      # EveOnline::ESI::Models::Bloodline.new(response.first) # => assets
+      # EveOnline::ESI::Models::Bloodline.new(response.first) # => bloodline
       #
       expect(EveOnline::ESI::Models::Bloodline).to receive(:new).with(response.first).and_return(bloodline)
     end
@@ -59,7 +59,7 @@ describe EveOnline::ESI::Bloodlines do
 
   describe '#url' do
     specify do
-      expect(subject.url).to eq(described_class::API_ENDPOINT)
+      expect(subject.url).to eq('https://esi.tech.ccp.is/v1/universe/bloodlines/?datasource=tranquility&language=en-us')
     end
   end
 end